Update Signed methods and fix a bug
Signed methods use their unsigned counterparts as the binary representation is identical and it isolates the logic to a single method.
A bug is fixed where a signed byte is cast to an integer and packed into a number with the bytes from the extension affecting the result. This also resolves the issue of what to do when an integer larger than a byte is fed into a method.